move chainbase related functions from Manager to ChainbaseManager.
