+++ framework/src/main/java/org/tron/core/db/accountstate/callback/AccountStateCallBack.java
@@ -10,9 +10,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.tron.common.crypto.Hash;
 import org.tron.common.utils.ByteUtil;
+import org.tron.core.ChainBaseManager;
 import org.tron.core.capsule.BlockCapsule;
 import org.tron.core.capsule.utils.RLP;
-import org.tron.core.db.Manager;
 import org.tron.core.db.accountstate.AccountStateCallBackUtils;
 import org.tron.core.db.accountstate.storetrie.AccountStateStoreTrie;
 import org.tron.core.exception.BadBlockException;
@@ -27,7 +27,7 @@ public class AccountStateCallBack extends AccountStateCallBackUtils {
   private TrieImpl trie;
 
   @Setter
-  private Manager manager;
+  private ChainBaseManager chainBaseManager;
 
   @Autowired
   private AccountStateStoreTrie db;
@@ -53,13 +53,14 @@ public class AccountStateCallBack extends AccountStateCallBackUtils {
   public void preExecute(BlockCapsule blockCapsule) {
     this.blockCapsule = blockCapsule;
     this.execute = true;
-    this.allowGenerateRoot = manager.getDynamicPropertiesStore().allowAccountStateRoot();
+    this.allowGenerateRoot = chainBaseManager.getDynamicPropertiesStore().allowAccountStateRoot();
     if (!exe()) {
       return;
     }
     byte[] rootHash = null;
     try {
-      BlockCapsule parentBlockCapsule = manager.getBlockById(blockCapsule.getParentBlockId());
+      BlockCapsule parentBlockCapsule =
+          chainBaseManager.getBlockById(blockCapsule.getParentBlockId());
       rootHash = parentBlockCapsule.getInstance().getBlockHeader().getRawData()
           .getAccountStateRoot().toByteArray();
     } catch (Exception e) {
